new cursors and new cell creating rool
There are two new cursors for brush and fill instrument. The classical cursor changes to one of new ones when it is on 2D-field. New rool: if you press ctrl+RMB to create a new cell, new cell will be created next to clicked cell. If you only press RMB and cell types are different then cell just changes as it was in previous version.